+++ src/activities/settings/OtaUpdateActivity.h
@@ -20,12 +20,14 @@ class OtaUpdateActivity : public Activity {
   // Can't initialize this to 0 or the first render doesn't happen
   static constexpr unsigned int UNINITIALIZED_PERCENTAGE = 111;
 
+  bool radioOwned = false;
   State state = WIFI_SELECTION;
   ScreenTransitionRefresh screenTransitionRefresh;
   unsigned int lastUpdaterPercentage = UNINITIALIZED_PERCENTAGE;
   StrId failureMessage = StrId::STR_UPDATE_FAILED;
   OtaUpdater updater;
 
+  bool requireStation();
   void onWifiSelectionComplete(bool success);
   void runUpdateInstall();